+++ PlantUmlViewer/PlantUmlViewer/Forms/PreviewWindow.cs
@@ -7,6 +7,11 @@ using System.Windows.Forms;
 using System.Runtime.InteropServices;
 using System.Diagnostics;
 using System.Text.RegularExpressions;
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using System.Linq;
+using System.Text;
+using System.Drawing.Imaging;
 
 using Kbg.NppPluginNET.PluginInfrastructure;
 
@@ -28,11 +33,60 @@ namespace PlantUmlViewer.Forms
         private readonly Func<string> getText;
         private readonly SettingsService settings;
 
+        private bool? isLight;
         private Color colorSuccess;
         private Color colorFailure;
 
-        private SvgDocument svgImage;
         private CancellationTokenSource refreshCancellationTokenSource;
+
+        #region Images
+        private object imagesLock = new object();
+        private int selectedImageIndex = 0;
+        private ReadOnlyCollection<SvgDocument> svgImages;
+
+        private void UpdateImages(IEnumerable<SvgDocument> newImages)
+        {
+            lock (imagesLock)
+            {
+                //Update the images
+                List<SvgDocument> newImagesList = newImages.ToList();
+                svgImages = new ReadOnlyCollection<SvgDocument>(newImagesList);
+                //Update the text of the selected diagram and visibility
+                SetSelectedImage(Math.Min(selectedImageIndex, svgImages.Count - 1));
+            }
+        }
+
+        private int GetSelectedImageIndex()
+        {
+            lock (imagesLock)
+            {
+                return selectedImageIndex;
+            }
+        }
+
+        private SvgDocument GetSelectedImage()
+        {
+            lock (imagesLock)
+            {
+                return svgImages[selectedImageIndex];
+            }
+        }
+
+        private void SetSelectedImage(int index)
+        {
+            Debug.WriteLine($"Selecting image at index {index}", nameof(PreviewWindow));
+            lock (imagesLock)
+            {
+                selectedImageIndex = index;
+                label_SelectedDiagram.Visible = svgImages.Count > 1;
+                label_SelectedDiagram.Text = (selectedImageIndex + 1).ToString();
+                tableLayoutPanel_Navigation.Visible = svgImages.Count > 1;
+                button_NextDiagram.Enabled = selectedImageIndex < svgImages.Count - 1;
+                button_PreviousDiagram.Enabled = selectedImageIndex > 0;
+                imageBox_Diagram.Image = GetDiagramImage(1);
+            }
+        }
+        #endregion Images
 
         public event EventHandler<EventArgs> DockablePanelClose;
 
@@ -45,6 +99,15 @@ namespace PlantUmlViewer.Forms
 
             InitializeComponent();
 
+            //Add some tool tips
+            toolTip_Buttons.SetToolTip(button_Refresh, "Refresh");
+            toolTip_Buttons.SetToolTip(button_Export, "Export");
+            toolTip_Buttons.SetToolTip(button_ZoomIn, "Zoom in");
+            toolTip_Buttons.SetToolTip(button_ZoomOut, "Zoom out");
+            toolTip_Buttons.SetToolTip(button_PreviousDiagram, "Previous diagram");
+            toolTip_Buttons.SetToolTip(button_NextDiagram, "Next diagram");
+
+            //Handle zoom changes
             imageBox_Diagram.ZoomChanged += ImageBox_ZoomChanged;
             ImageBox_ZoomChanged(this, null);
         }
@@ -64,11 +127,76 @@ namespace PlantUmlViewer.Forms
             base.WndProc(ref m);
         }
 
-        private void ImageBox_ZoomChanged(object sender, EventArgs e)
+        public void SetStyle(Color editorBackgroundColor)
         {
-            toolStripStatusLabel_Zoom.Text = $"{imageBox_Diagram.Zoom}%";
+            //Check the current style and update if necessary
+            bool newIsLight = editorBackgroundColor.GetBrightness() > 0.4;
+            if (isLight == newIsLight)
+            {
+                return;
+            }
+            isLight = newIsLight;
+            Debug.WriteLine("Setting style", nameof(PreviewWindow));
+
+            imageBox_Diagram.BackColor = editorBackgroundColor;
+
+            Color buttonBackColor;
+            Color buttonForeColor;
+            if (isLight == true)
+            {
+                //Light
+                colorSuccess = Color.LightGreen;
+                colorFailure = Color.Tomato;
+                BackColor = SystemColors.Control;
+                buttonBackColor = SystemColors.Control;
+                buttonForeColor = SystemColors.ControlText;
+                label_SelectedDiagram.ForeColor = SystemColors.ControlText;
+                loadingCircleToolStripMenuItem_Refreshing.LoadingCircleControl.Color = Color.DarkGray;
+                statusStrip_Bottom.BackColor = SystemColors.Control;
+                statusStrip_Bottom.ForeColor = SystemColors.ControlText;
+            }
+            else
+            {
+                //Dark
+                colorSuccess = Color.DarkGreen;
+                colorFailure = Color.DarkRed;
+                BackColor = SystemColors.ControlDarkDark;
+                buttonBackColor = SystemColors.ControlDarkDark;
+                buttonForeColor = SystemColors.ControlLightLight;
+                label_SelectedDiagram.ForeColor = SystemColors.ControlLightLight;
+                loadingCircleToolStripMenuItem_Refreshing.LoadingCircleControl.Color = Color.LightGray;
+                statusStrip_Bottom.BackColor = SystemColors.ControlDarkDark;
+                statusStrip_Bottom.ForeColor = SystemColors.ControlLightLight;
+            }
+            ColorMap[] buttonImageColorMap = new ColorMap[] {
+                new ColorMap()
+                {
+                    OldColor = Color.Black,
+                    NewColor = buttonForeColor
+                }
+            };
+
+            button_Refresh.BackColor = buttonBackColor;
+            button_Refresh.ForeColor = buttonForeColor;
+            button_Refresh.BackgroundImage = RemapImage(Resources.Refresh, buttonImageColorMap);
+            button_Export.BackColor = buttonBackColor;
+            button_Export.ForeColor = buttonForeColor;
+            button_Export.BackgroundImage = RemapImage(Resources.Save, buttonImageColorMap);
+            button_ZoomIn.BackColor = buttonBackColor;
+            button_ZoomIn.ForeColor = buttonForeColor;
+            button_ZoomIn.BackgroundImage = RemapImage(Resources.ZoomIn, buttonImageColorMap);
+            button_ZoomOut.BackColor = buttonBackColor;
+            button_ZoomOut.ForeColor = buttonForeColor;
+            button_ZoomOut.BackgroundImage = RemapImage(Resources.ZoomOut, buttonImageColorMap);
+            button_PreviousDiagram.BackColor = buttonBackColor;
+            button_PreviousDiagram.ForeColor = buttonForeColor;
+            button_PreviousDiagram.BackgroundImage = RemapImage(Resources.Previous, buttonImageColorMap);
+            button_NextDiagram.BackColor = buttonBackColor;
+            button_NextDiagram.ForeColor = buttonForeColor;
+            button_NextDiagram.BackgroundImage = RemapImage(Resources.Next, buttonImageColorMap);
         }
 
+        #region Button events
         public async void Button_Refresh_Click(object sender, EventArgs e)
         {
             //Cancel if already running
@@ -83,7 +211,6 @@ namespace PlantUmlViewer.Forms
             {
                 loadingCircleToolStripMenuItem_Refreshing.LoadingCircleControl.Active = true;
                 loadingCircleToolStripMenuItem_Refreshing.Visible = true;
-                button_Refresh.Text = "Cancel";
 
                 RendererFactory factory = new RendererFactory();
                 IPlantUmlRenderer renderer = factory.CreateRenderer(new PlantUmlSettings()
@@ -93,12 +220,15 @@ namespace PlantUmlViewer.Forms
                     RenderingMode = RenderingMode.Local
                 });
 
+                //The response could contain multiple XML documents / SVG images
+                List<SvgDocument> images = new List<SvgDocument>();
+
                 text = getText();
                 if (string.IsNullOrWhiteSpace(text))
                 {
-                    using (MemoryStream stream = new MemoryStream(Resources.Empty))
+                    using (MemoryStream memoryStream = new MemoryStream(Resources.Empty))
                     {
-                        svgImage = SvgDocument.Open<SvgDocument>(stream);
+                        images.Add(SvgDocument.Open<SvgDocument>(memoryStream));
                     }
                 }
                 else
@@ -106,18 +236,29 @@ namespace PlantUmlViewer.Forms
                     refreshCancellationTokenSource = new CancellationTokenSource();
                     byte[] bytes = await renderer.RenderAsync(text, OutputFormat.Svg,
                         refreshCancellationTokenSource.Token).ConfigureAwait(true);
-                    using (MemoryStream ms = new MemoryStream(bytes))
+
+                    //Find all start XML declarations to parse multiple images
+                    List<int> xmlStartDeclararationIndices = PatternAt(bytes, Encoding.UTF8.GetBytes("<?xml")).ToList();
+                    for (int i = 0; i < xmlStartDeclararationIndices.Count; i++)
                     {
-                        svgImage = SvgDocument.Open<SvgDocument>(ms);
+                        int start = xmlStartDeclararationIndices[i];
+                        int end = i == xmlStartDeclararationIndices.Count - 1 ? bytes.Length : xmlStartDeclararationIndices[i + 1];
+                        using (MemoryStream memoryStream = new MemoryStream(bytes, start, end - start))
+                        {
+                            images.Add(SvgDocument.Open<SvgDocument>(memoryStream));
+                        }
                     }
                 }
+                Debug.WriteLine($"{images.Count} image(s) generated", nameof(PreviewWindow));
 
                 this.InvokeIfRequired(() =>
                 {
-                    imageBox_Diagram.Image = GetDiagramImage(1);
+                    UpdateImages(images);
                     toolStripStatusLabel_Time.Text = DateTime.Now.ToShortTimeString();
                     toolStripStatusLabel_Time.BackColor = colorSuccess;
                     button_Export.Enabled = true;
+                    button_ZoomIn.Enabled = true;
+                    button_ZoomOut.Enabled = true;
                     ToolStripMenuItem_Diagram_ExportFile.Enabled = true;
                     ToolStripMenuItem_Diagram_CopyToClipboard.Enabled = true;
                 });
@@ -166,7 +307,6 @@ namespace PlantUmlViewer.Forms
                 {
                     loadingCircleToolStripMenuItem_Refreshing.Visible = false;
                     loadingCircleToolStripMenuItem_Refreshing.LoadingCircleControl.Active = false;
-                    button_Refresh.Text = "Refresh";
                     refreshCancellationTokenSource = null;
                 });
             }
@@ -179,7 +319,7 @@ namespace PlantUmlViewer.Forms
                 using (SaveFileDialog saveFileDialog = new SaveFileDialog()
                 {
                     Filter = "PNG file|*.png|SVG file|*.svg",
-                    FileName = $"{Path.GetFileNameWithoutExtension(getFilePath())}.png",
+                    FileName = $"{Path.GetFileNameWithoutExtension(getFilePath())}{(svgImages.Count > 1 ? $"_{GetSelectedImageIndex() + 1}" : "")}.png",
                     InitialDirectory = Path.GetDirectoryName(getFilePath())
                 })
                 {
@@ -191,7 +331,7 @@ namespace PlantUmlViewer.Forms
                                 GetDiagramImage(settings.Settings.ExportSizeFactor).Save(saveFileDialog.FileName);
                                 break;
                             case ".svg":
-                                svgImage.Write(saveFileDialog.FileName);
+                                GetSelectedImage().Write(saveFileDialog.FileName);
                                 break;
                             default:
                                 throw new Exception("Invalid file extension");
@@ -205,23 +345,51 @@ namespace PlantUmlViewer.Forms
             }
         }
 
+        private void Button_ZoomIn_Click(object sender, EventArgs e)
+        {
+            imageBox_Diagram.ZoomIn();
+        }
+
+        private void Button_ZoomOut_Click(object sender, EventArgs e)
+        {
+            imageBox_Diagram.ZoomOut();
+        }
+
+        private void Button_PreviousDiagram_Click(object sender, EventArgs e)
+        {
+            SetSelectedImage(GetSelectedImageIndex() - 1);
+        }
+
+        private void Button_NextDiagram_Click(object sender, EventArgs e)
+        {
+            SetSelectedImage(GetSelectedImageIndex() + 1);
+        }
+        #endregion Button events
+
         private void ToolStripMenuItem_Diagram_CopyToClipboard_Click(object sender, EventArgs e)
         {
             Clipboard.SetImage(GetDiagramImage(settings.Settings.ExportSizeFactor));
         }
 
+        private void ImageBox_ZoomChanged(object sender, EventArgs e)
+        {
+            toolStripStatusLabel_Zoom.Text = $"{imageBox_Diagram.Zoom}%";
+        }
+
         private Image GetDiagramImage(decimal exportSizeFactor)
         {
+            SvgDocument selectedImage = GetSelectedImage();
+
             //Resize (See: https://github.com/svg-net/SVG/blob/master/Source/SvgDocument.Drawing.cs#L217)
-            SizeF svgSize = svgImage.GetDimensions();
+            SizeF svgSize = selectedImage.GetDimensions();
             SizeF imageSize = svgSize;
-            svgImage.RasterizeDimensions(ref imageSize,
+            selectedImage.RasterizeDimensions(ref imageSize,
                 (int)Math.Round((decimal)svgSize.Width * exportSizeFactor), (int)Math.Round((decimal)svgSize.Height * exportSizeFactor));
             Size bitmapSize = Size.Round(imageSize);
             Bitmap image = new Bitmap(bitmapSize.Width, bitmapSize.Height);
 
             //Set background if defined in SVG
-            if (svgImage.TryGetAttribute("background", out string backgroundAttribute))
+            if (selectedImage.TryGetAttribute("background", out string backgroundAttribute))
             {
                 using (Graphics g = Graphics.FromImage(image))
                 using (SolidBrush brush = new SolidBrush(ColorTranslator.FromHtml(backgroundAttribute)))
@@ -231,65 +399,54 @@ namespace PlantUmlViewer.Forms
             }
 
             //Render
-            using (var renderer = SvgRenderer.FromImage(image))
+            using (ISvgRenderer renderer = SvgRenderer.FromImage(image))
             {
                 renderer.ScaleTransform(imageSize.Width / svgSize.Width, imageSize.Height / svgSize.Height);
-                svgImage.Draw(renderer);
+                selectedImage.Draw(renderer);
             }
             return image;
         }
 
-        public void SetStyle(Color editorBackgroundColor)
+        private static Bitmap RemapImage(Bitmap image, ColorMap[] colorMap)
         {
-            imageBox_Diagram.BackColor = editorBackgroundColor;
-
-            Color buttonBackColor;
-            Color buttonForeColor;
-            if (editorBackgroundColor.GetBrightness() > 0.4)
+            using (ImageAttributes imageAttributes = new ImageAttributes())
             {
-                //Light
-                colorSuccess = Color.LightGreen;
-                colorFailure = Color.Tomato;
-                BackColor = SystemColors.Control;
-                buttonBackColor = SystemColors.Control;
-                buttonForeColor = SystemColors.ControlText;
-                loadingCircleToolStripMenuItem_Refreshing.LoadingCircleControl.Color = Color.DarkGray;
-                statusStrip_Bottom.BackColor = SystemColors.Control;
-                statusStrip_Bottom.ForeColor = SystemColors.ControlText;
+                imageAttributes.SetRemapTable(colorMap);
+                Bitmap newImage = new Bitmap(image.Width, image.Height);
+                using (Graphics g = Graphics.FromImage(newImage))
+                {
+                    g.DrawImage(image, new Rectangle(0, 0, image.Width, image.Height), 0, 0,
+                        image.Width, image.Height, GraphicsUnit.Pixel, imageAttributes);
+                }
+                return newImage;
             }
-            else
-            {
-                //Dark
-                colorSuccess = Color.DarkGreen;
-                colorFailure = Color.DarkRed;
-                BackColor = SystemColors.ControlDarkDark;
-                buttonBackColor = SystemColors.ControlDarkDark;
-                buttonForeColor = SystemColors.ControlLightLight;
-                loadingCircleToolStripMenuItem_Refreshing.LoadingCircleControl.Color = Color.LightGray;
-                statusStrip_Bottom.BackColor = SystemColors.ControlDarkDark;
-                statusStrip_Bottom.ForeColor = SystemColors.ControlLightLight;
-            }
-
-            button_Export.BackColor = buttonBackColor;
-            button_Export.ForeColor = buttonForeColor;
-            button_Refresh.BackColor = buttonBackColor;
-            button_Refresh.ForeColor = buttonForeColor;
         }
 
         private static string ReadLine(string text, int lineNumber)
         {
-            var reader = new StringReader(text);
-
-            string line;
-            int currentLineNumber = 0;
-            do
+            using (StringReader reader = new StringReader(text))
             {
-                currentLineNumber++;
-                line = reader.ReadLine();
+                string line;
+                int currentLineNumber = 0;
+                do
+                {
+                    currentLineNumber++;
+                    line = reader.ReadLine();
+                }
+                while (line != null && currentLineNumber < lineNumber);
+                return (currentLineNumber == lineNumber) ? line : "";
             }
-            while (line != null && currentLineNumber < lineNumber);
+        }
 
-            return (currentLineNumber == lineNumber) ? line : "";
+        private static IEnumerable<int> PatternAt(byte[] source, byte[] pattern)
+        {
+            for (int i = 0; i < source.Length; i++)
+            {
+                if (source.Skip(i).Take(pattern.Length).SequenceEqual(pattern))
+                {
+                    yield return i;
+                }
+            }
         }
     }
 }