+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # X.X.X.X (XX.XX.XXXX)
+- Added support for multiple diagrams per document
+- Fixed using the text of the selected document in split mode
+- Used icons for all buttons
 - Used loading circle instead of progress bar to show refreshing
 - Added button to check for update in about window
-- Fixed using the text of the selected document in split mode
 
 # 1.2.0.6 (11.06.2022)
 - Added export of SVG files